fix size
needs to be same data type as counters

+++ pcsx2/SPU2/Voice.h
@@ -351,7 +351,7 @@ namespace SPU
 			m_Wpos += n;
 		}
 
-		size_t Size() { return m_Wpos - m_Rpos; }
+		u16 Size() { return static_cast<u16>(m_Wpos - m_Rpos); }
 
 		s16* Get()
 		{
